🎨 Refactor WS design
