+++ fastcrawler_ui/controllers/ws.py
@@ -1,74 +1,18 @@
-from fastapi import WebSocket
-from pydantic import BaseModel
+import asyncio
+
+from fastapi import Depends, WebSocket, WebSocketDisconnect
+
+from fastcrawler_ui.repository.ws import ConnectionRepository
 
 
-class Message(BaseModel):
-    """
-    Represents a message sent over WebSocket.
-    """
+class WSController:
+    def __init__(self, ws_connection_repo: ConnectionRepository = Depends(ConnectionRepository)):
+        self.ws_connection_repo = ws_connection_repo
 
-    content: str
-    sender: str
-
-
-class ConnectionManager:
-    """
-    Manages WebSocket connections for the FastAPI application.
-    """
-
-    def __init__(self):
-        """
-        Initializes the ConnectionManager.
-        """
-        self.active_connections: set[WebSocket] = set()
-
-    async def connect(self, websocket: WebSocket):
-        """
-        Handles the connection of a new WebSocket client.
-
-        Args:
-            websocket (WebSocket): The WebSocket instance representing the client connection.
-
-        Returns:
-            None
-        """
-        await websocket.accept()
-        self.active_connections.add(websocket)
-
-    def disconnect(self, websocket: WebSocket):
-        """
-        Handles the disconnection of a WebSocket client.
-
-        Args:
-            websocket (WebSocket): The WebSocket instance representing the client connection.
-
-        Returns:
-            None
-        """
-        self.active_connections.remove(websocket)
-
-    async def send_personal_message(self, message: Message, websocket: WebSocket):
-        """
-        Sends a personal message to a specific WebSocket client.
-
-        Args:
-            message (Message): The message to send.
-            websocket (WebSocket): The WebSocket instance representing the client connection.
-
-        Returns:
-            None
-        """
-        await websocket.send_json(message.model_dump())
-
-    async def broadcast(self, message: Message):
-        """
-        Broadcasts a message to all connected WebSocket clients.
-
-        Args:
-            message (Message): The message to broadcast.
-
-        Returns:
-            None
-        """
-        for connection in self.active_connections:
-            await connection.send_json(message.model_dump())
+    async def get_connection(self, websocket: WebSocket):
+        await self.ws_connection_repo.connect(websocket)
+        try:
+            while True:
+                await asyncio.sleep(1)
+        except WebSocketDisconnect:
+            self.ws_connection_repo.disconnect(websocket)
